backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
pymongo==4.6.0
espn-api==0.37.1
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
requests==2.32.3
urllib3==2.2.2
serverless-wsgi==3.0.4
aws-wsgi
aws-serverless-express
zappa==0.54.1
